implement new client-system
