+++ shepherd_core/shepherd_core/__init__.py
@@ -19,8 +19,8 @@ from .logger import get_verbose_level
 from .logger import increase_verbose_level
 from .logger import logger
 from .reader import Reader
-from .testbed_client.client import TestbedClient
-from .testbed_client.client import tb_client
+from .testbed_client import tb_client
+from .testbed_client.client_web import WebClient
 from .version import version
 from .writer import Writer
 
@@ -41,7 +41,7 @@ __all__ = [
     "local_now",
     "Calc_t",
     "Compression",
-    "TestbedClient",
+    "WebClient",
     "tb_client",  # using this (instead of the Class) is the cleaner, but less pythonic way
     "Inventory",
 ]
