+++ shepherd_core/shepherd_core/testbed_client/__init__.py
@@ -1,7 +1,9 @@
 """Client to access a testbed-instance for controlling experiments."""
 
-from .client import tb_client
+from .client_fixtures import FixturesClient
 from .user_model import User
+
+tb_client = FixturesClient()
 
 __all__ = [
     "tb_client",
